Lógica do Programa

+++ script.js
@@ -6,8 +6,8 @@ var product3Message = document.getElementById("product3");
 
 calcBtn.addEventListener("click", calcular);
 function calcular() {
-  var product3 = parseInt(price.value) * 0.5;
-  var promoPrice = parseInt(price.value) * 3 - product3;
+  var product3 = parseFloat(price.value) * 0.5;
+  var promoPrice = parseFloat(price.value) * 3 - product3;
   promo.textContent =
     product.value.toUpperCase() +
     " - Promoção: Leve 3 por R$: " +
